fix(api): JSON-number serialization for all Decimal-backed response fields (e2e-sweep ISSUE-011)
Class sweep of backend/schemas/**: every Pydantic response model field backed
by a SQLAlchemy Numeric column (or otherwise carrying Decimal at runtime) now
serializes as a JSON number, not a Decimal-as-string. Root cause confirmed
end-to-end through FastAPI's actual response path: Pydantic v2 stringifies a
field typed Decimal on model_dump_json/model_dump(mode="json"); a field typed
float correctly coerces a Decimal ORM value via from_attributes. Applies the
#145 mechanism (cast Decimal -> float) at the schema field-type level.

Fixed (Decimal -> float) in: WIPHoldResponse, WIPAgingResponse,
WIPAgingAdjustedResponse, TotalHoldDurationResponse (hold.py);
EmployeeLineAssignmentResponse; JobResponse (overridden, JobBase/Create kept
Decimal); AttendanceRecordResponse, AbsenteeismCalculationResponse;
ShiftCoverageResponse; AvailabilityCalculationResponse; WorkOrderResponse,
WorkOrderWithMetrics; QualityInspectionResponse, PPMCalculationResponse,
DPMOCalculationResponse, FPYRTYCalculationResponse; ProductionEntryResponse,
ProductionEntryWithKPIs, KPICalculationResponse; AlertResponse,
AlertConfigResponse (overridden), OTDRiskAlert, QualityTrendAlert,
CapacityAlert. Create/Update input schemas keep Decimal for validation
precision (not returned to clients). Two field overrides (JobResponse,
AlertConfigResponse) narrow their base class's Decimal typing to float and
need `# type: ignore[assignment]` — a standard mypy/Pydantic limitation for
input/output type divergence on an inherited field.

Guard test (tests/test_models/test_decimal_response_serialization.py):
(1) a structural sweep over every response_model= class across backend/routes
asserting none exposes a Decimal field — regression backstop for future
additions; (2) 24 parametrized cases constructing each fixed model with real
Decimal input and asserting the JSON dump decodes to a number, not a string.
Runs on SQLite CI (the defect is Pydantic-serialization-level, not
driver-level).

Caught and fixed 2 pre-existing tests
(test_employee_line_assignment_routes.py) that asserted the broken
stringified-Decimal contract directly — a live instance of the same
green-while-dead pattern the #149 e2e guard closed for report buttons.

+++ backend/schemas/attendance.py
@@ -122,9 +122,9 @@ class AttendanceRecordResponse(BaseModel):
     line_id: Optional[int] = None
     shift_date: datetime
     shift_id: Optional[int] = None
-    scheduled_hours: Decimal
-    actual_hours: Optional[Decimal] = None
-    absence_hours: Optional[Decimal] = None
+    scheduled_hours: float
+    actual_hours: Optional[float] = None
+    absence_hours: Optional[float] = None
     is_absent: int
     absence_type: Optional[str] = None
     covered_by_employee_id: Optional[int] = None
@@ -167,10 +167,10 @@ class AbsenteeismCalculationResponse(BaseModel):
     shift_id: int
     start_date: date
     end_date: date
-    total_scheduled_hours: Decimal
-    total_hours_worked: Decimal
-    total_hours_absent: Decimal
-    absenteeism_rate: Decimal
+    total_scheduled_hours: float
+    total_hours_worked: float
+    total_hours_absent: float
+    absenteeism_rate: float
     total_employees: int
     total_absences: int
     calculation_timestamp: datetime
